[FIX] Remove duplicate sidebar logout, layout borders and standardize navigation labels
