+++ src/test/AdminLayout.test.tsx
@@ -15,13 +15,12 @@ describe("AdminLayout", () => {
       </MemoryRouter>,
     );
 
-    expect(screen.getAllByText("Dashboard")[0]).toBeInTheDocument();
+    expect(screen.getAllByText("Tableau de Bord")[0]).toBeInTheDocument();
     expect(screen.getByText("nested-dashboard")).toBeInTheDocument();
 
     const aside = screen.getByRole("complementary");
     fireEvent.mouseEnter(aside);
-    expect(screen.getByText("Tableau de Bord")).toBeInTheDocument();
-    expect(screen.getByText("Administrateur")).toBeInTheDocument();
+    expect(screen.getAllByText("Tableau de Bord")[0]).toBeInTheDocument();
   });
 
   it("handles image error in Sidebar", () => {
@@ -34,32 +33,6 @@ describe("AdminLayout", () => {
     const logo = screen.getByAltText("Aegis AI Logo");
     fireEvent.error(logo);
     expect(logo).toHaveAttribute("src", "/logo.png");
-  });
-
-  it("performs logout from Sidebar", async () => {
-    render(
-      <MemoryRouter initialEntries={["/"]}>
-        <Routes>
-          <Route element={<AdminLayout />}>
-            <Route path="/" element={<div>Dashboard</div>} />
-          </Route>
-          <Route path="/login" element={<div>LoginPage</div>} />
-        </Routes>
-      </MemoryRouter>,
-    );
-
-    // Sidebar logout button is visible when expanded or in the bottom when collapsed
-    // Let's expand it first
-    const aside = screen.getByRole("complementary");
-    fireEvent.mouseEnter(aside);
-
-    const logoutButtons = screen.getAllByTitle("Déconnexion");
-    fireEvent.click(logoutButtons[0]);
-
-    expect(screen.getByText("Se déconnecter")).toBeInTheDocument();
-    fireEvent.click(screen.getByText("Se déconnecter"));
-
-    expect(screen.getByText("LoginPage")).toBeInTheDocument();
   });
 
   it("performs logout from Topbar", async () => {
@@ -78,7 +51,9 @@ describe("AdminLayout", () => {
     const logoutButton = topbar.querySelector('button[title="Déconnexion"]');
     fireEvent.click(logoutButton!);
 
+    expect(screen.getByText("Se déconnecter")).toBeInTheDocument();
     fireEvent.click(screen.getByText("Se déconnecter"));
+
     expect(screen.getByText("LoginPage")).toBeInTheDocument();
   });
 });
